crud cajero

+++ src/main/java/com/elmanalubricentro/ElMana/cajero/controller/CajeroController.java
@@ -1,8 +1,14 @@
 package com.elmanalubricentro.ElMana.cajero.controller;
 
+import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
+import org.apache.catalina.connector.Response;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,41 +17,65 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
 import com.elmanalubricentro.ElMana.cajero.entity.Cajero;
-import com.elmanalubricentro.ElMana.cajero.service.ICajeroService;
+import com.elmanalubricentro.ElMana.cajero.service.CajeroService;
 
 @RestController
 @RequestMapping("/cajero")
 public class CajeroController {
 
     @Autowired
-    private ICajeroService cajeroService;
+    private CajeroService cajeroService;
 
     @PostMapping("/crear")
-    public String crearCajero(@RequestBody Cajero cajero) {
-        cajeroService.saveCajero(cajero);
+    ResponseEntity<?> createCajero(@RequestBody Cajero cajero) {
+        Cajero newCajero = cajeroService.saveCajero(cajero);
+
+        URI location = ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(cajero.getId_cajero())
+                .toUri();
         
-        return "Cajero creado correctamente";
+        return ResponseEntity.created(location).body(newCajero);
     }
 
-    @GetMapping("/traer")
-    public List<Cajero> traerCajeros() {
-        return cajeroService.getCajeros();
+    @GetMapping("/lista")
+    public ResponseEntity<?> getAllCajeros() {
+        try {
+            List<Cajero> cajeros = cajeroService.getCajeros();
+            return ResponseEntity.ok(cajeros);
+
+        } catch (DataAccessException e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body("Error al acceder a la base de datos");
+
+        } catch(Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .body("No se pudo conectar al servidor");
+        }
     }
 
     @PutMapping("/editar")
-    public String editCajero(@RequestBody Cajero cajero) {
+    public ResponseEntity<?> editCajero(@RequestBody Cajero cajero) {
         cajeroService.editCajero(cajero);
 
-        return "Cajero editado correctamente";
+        Cajero updatedCajero = cajeroService.editCajero(cajero);
+        return ResponseEntity.ok(updatedCajero);
     }
        
     @DeleteMapping("/borrar/{id}")
-    public String deleteCajero(@PathVariable Long id) {
-        cajeroService.deleteCajero(id);
+    public ResponseEntity<?> deleteCajero(@PathVariable Long id) {
+        Optional <Cajero> cajero = cajeroService.getById(id);
+        
+        if(cajero.isPresent() && cajero.get().isActivo()) {
+            cajeroService.deleteCajero(id);
+            return ResponseEntity.ok().build();
+        }
 
-        return "Cajero eliminado correctamente";
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cliente no encontrado");
     }
 
     @GetMapping("/helloWorld")
